Only show the appropriate selection according to selection mode
Firstly, disable any shadow casting for the vertices, faces and edges. I
think Godot 3.0 had `cast_shadows` default to off, which is probably why
it wasn't explictly set in code. Godot 4.0 however has it default to on.

Secondly, this will only show the corresponding selection the the current
selection mode. When the mode is VERTEX we only show vertices, for FACE
we only show faces. However i choose to make sure edges are always
visible because it's always needed to give you an idea of how the mesh
is structured. Also, vertices and faces are hidden by default to help
reduce visual clutter when selecting a new PlyEditor.